Integrate tools/kern_table.py into tools/build.py
Build legacy kerning by default.
