+++ Makefile
@@ -6,15 +6,12 @@ ALL_FONTS=$(CORE_FONTS) C L LI M MI
 
 MASTER_DIR=master_ttf
 BUILD_DIR=build
-LEGACY_KERN_DIR=$(BUILD_DIR)/legacy_kern
 
 TTF=$(ALL_FONTS:%=$(BUILD_DIR)/$(NORMAL_NAME)-%.ttf)
 MONO_TTF=$(CORE_FONTS:%=$(BUILD_DIR)/$(MONO_NAME)-%.ttf)
 
 VTT_TTF=$(ALL_FONTS:%=$(SRC_DIR)/$(NORMAL_NAME)-%.ttf)
 VTT_MONO_TTF=$(CORE_FONTS:%=$(SRC_DIR)/$(MONO_NAME)-%.ttf)
-
-LEGACY_KERN_TTF=$(ALL_FONTS:%=$(LEGACY_KERN_DIR)/$(NORMAL_NAME)-%.ttf)
 
 
 all: ttf
@@ -35,11 +32,6 @@ $(SRC_DIR)/%.ttf: $(SRC_DIR)/%.ufo $(SRC_DIR)/%.ufo/*.plist \
 $(BUILD_DIR)/%.ttf: $(SRC_DIR)/%.ttf
 	@mkdir -p $(BUILD_DIR)
 	@python -m vttLib compile --ship $< $@
-
-
-$(LEGACY_KERN_DIR)/%.ttf: $(SRC_DIR)/%.ufo $(BUILD_DIR)/%.ttf
-		@mkdir -p $(LEGACY_KERN_DIR)
-		@python tools/kern_table.py $< $(word 2,$^) $@
 
 clean:
 	@rm -rf $(BUILD_DIR)
